+++ demo/src/main/java/com/example/demo/repository/CountryRepository.java
@@ -3,5 +3,5 @@ package com.example.demo.repository;
 import com.example.demo.model.Country;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CountryRepository extends JpaRepository<Country, Long> {
+public interface CountryRepository extends JpaRepository<Country, Integer> {
 }
